폼 목록 페이지 리사이클러뷰 + 프래그먼트 + 레트로핏 (미확인)

+++ app/src/main/java/com/example/umc_hackathon/post/PostRetrofitInterface.kt
@@ -7,7 +7,5 @@ import retrofit2.http.Query
 
 interface PostRetrofitInterface {
     @GET("/posts")
-    fun getPostList(
-        @Query("categoryId") category: Int
-    ): Call<PostListResponse>
+    fun getPostList(@Query("categoryId") category: Long): Call<PostListResponse>
 }
